requirements.txt
PyYAML>=6.0
numpy>=1.21.6
pandas>=1.3.5
scikit-learn==1.2.2
scikit-surprise
myfm